Refactor onetime api
